feat(ui): prompt user to install provider if they want to create a workspace without having (the correct) one installed
