+++ desktop/src-tauri/tauri.conf.json
@@ -96,8 +96,8 @@
         "title": "Devpod Desktop",
         "width": 1200,
         "height": 800,
-        "minWidth": 800,
-        "minHeight": 600,
+        "minWidth": 1000,
+        "minHeight": 700,
         "x": 0,
         "y": 0,
         "focus": false,
