+++ desktop/src/main.tsx
@@ -9,7 +9,6 @@ import "xterm/css/xterm.css"
 import { DevPodProvider, SettingsProvider } from "./contexts"
 import { router } from "./routes"
 import { ThemeProvider } from "./Theme"
-import { LightMode } from "@chakra-ui/react"
 
 dayjs.extend(relativeTime)
 
@@ -40,20 +39,18 @@ function Root() {
   return (
     <StrictMode>
       <ThemeProvider>
-        <LightMode>
-          <QueryClientProvider client={queryClient}>
-            <SettingsProvider>
-              <DevPodProvider>
-                <RouterProvider router={router} />
-              </DevPodProvider>
-            </SettingsProvider>
-            {/* Will be disabled in production automatically */}
-            <ReactQueryDevtools
-              position="bottom-right"
-              toggleButtonProps={{ style: { margin: "0.5em 0.5em 2rem" } }}
-            />
-          </QueryClientProvider>
-        </LightMode>
+        <QueryClientProvider client={queryClient}>
+          <SettingsProvider>
+            <DevPodProvider>
+              <RouterProvider router={router} />
+            </DevPodProvider>
+          </SettingsProvider>
+          {/* Will be disabled in production automatically */}
+          <ReactQueryDevtools
+            position="bottom-right"
+            toggleButtonProps={{ style: { margin: "0.5em 0.5em 2rem" } }}
+          />
+        </QueryClientProvider>
       </ThemeProvider>
     </StrictMode>
   )
