--- desktop/src/views/Open/Open.tsx
@@ -1,3 +1,0 @@
-export function Open() {
-  return null
-}
