--- desktop/src/views/Open/index.ts
@@ -1,1 +1,0 @@
-export { Open } from "./Open"
