+++ desktop/src/views/Providers/AddProvider/index.ts
@@ -1,1 +1,2 @@
 export { AddProvider } from "./AddProvider"
+export { SetupProviderSteps } from "./SetupProviderSteps"
