+++ desktop/src/views/Providers/index.ts
@@ -1,4 +1,4 @@
-export { AddProvider } from "./AddProvider"
+export { AddProvider, SetupProviderSteps } from "./AddProvider"
 export { Provider } from "./Provider"
 export { Providers } from "./Providers"
 export { ListProviders } from "./ListProviders"
