+++ desktop/src/views/Workspaces/CreateWorkspace/constants.ts
@@ -1,0 +1,45 @@
+import {
+  CppSvg,
+  DotnetcorePng,
+  GoPng,
+  JavaPng,
+  NodejsPng,
+  PhpSvg,
+  PythonSvg,
+  RustSvg,
+} from "../../../images"
+
+export const WORKSPACE_EXAMPLES = [
+  {
+    image: PythonSvg,
+    source: "https://github.com/microsoft/vscode-remote-try-python",
+  },
+  {
+    image: NodejsPng,
+    source: "https://github.com/microsoft/vscode-remote-try-node",
+  },
+  {
+    image: GoPng,
+    source: "https://github.com/Microsoft/vscode-remote-try-go",
+  },
+  {
+    image: RustSvg,
+    source: "https://github.com/microsoft/vscode-remote-try-rust",
+  },
+  {
+    image: JavaPng,
+    source: "https://github.com/microsoft/vscode-remote-try-java",
+  },
+  {
+    image: PhpSvg,
+    source: "https://github.com/microsoft/vscode-remote-try-php",
+  },
+  {
+    image: CppSvg,
+    source: "https://github.com/microsoft/vscode-remote-try-cpp",
+  },
+  {
+    image: DotnetcorePng,
+    source: "https://github.com/microsoft/vscode-remote-try-dotnet",
+  },
+]
