+++ desktop/src/views/Workspaces/CreateWorkspace/index.ts
@@ -1,0 +1,1 @@
+export { CreateWorkspace } from "./CreateWorkspace"
