+++ desktop/src/views/Workspaces/CreateWorkspace/types.ts
@@ -1,4 +1,5 @@
-import { TProviderID } from "../../types"
+import { Routes } from "../../../routes"
+import { TProviderID, TWorkspaceID } from "../../../types"
 
 export const FieldName = {
   SOURCE: "source",
@@ -15,3 +16,14 @@ export type TFormValues = {
   [FieldName.ID]: string
   [FieldName.PREBUILD_REPOSITORY]: string
 }
+
+export type TCreateWorkspaceSearchParams = ReturnType<
+  typeof Routes.getWorkspaceCreateParamsFromSearchParams
+>
+export type TCreateWorkspaceArgs = Readonly<{
+  workspaceID: TWorkspaceID
+  providerID: TProviderID
+  prebuildRepositories: string[]
+  defaultIDE: string
+  workspaceSource: string
+}>
