+++ desktop/src/views/index.ts
@@ -1,5 +1,4 @@
 export * from "./Actions"
-export * from "./Open"
 export * from "./Providers"
 export * from "./Settings"
 export * from "./Workspaces"
